% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{circular}
\alias{circular}
\title{circular}
\usage{
circular(x, n)
}
\arguments{
\item{x}{Numeric variable.}

\item{n}{Number of rows to smooth over.}
}
\value{
Smoothed variable.
}
\description{
Smooth annual data using a circular smoother. (Shorthand for stats::filter because its syntax is very confusing.)
}
\examples{
\dontrun{


}

}
