% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan_set.R
\name{cmdstan_set}
\alias{cmdstan_set}
\title{cmdstan_set}
\usage{
cmdstan_set(user = NA)
}
\arguments{
\item{user}{OS username used for finding the cmdstan installation directory. Optional: By default, `Sys.getenv("USERPROFILE")` is called.}
}
\value{
No object -- fixes pathing issues for running cmdstan (if cmdstan is installed).
}
\description{
Sets cmdstan parameters to use cmdstan easily from {brms}. TODO: I'm not actually sure how necessary these fixes still are, but they were required mid-2024 to run cmdstan....
}
\examples{
\dontrun{
cmdstan_set()
}

}
