% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{count_na}
\usage{
count_na(x, pivot = FALSE, drop_zeros = TRUE)
}
\arguments{
\item{x}{: dataframe}

\item{pivot}{lgl : whether to pivot the output to long data formate}

\item{drop_zeros}{lgl : whether to drop groups or columns with zero NAs}
}
\value{
wide or long format dataframe with the count of zeros across all the columns in the original dataset
}
\description{
Check the number of NAs for all columns
}
\examples{
tibble::tibble(a = c(1, 2, NA),
               b = c(NA, NA, 'Q'),
               c = c(3, 2, 1)) \%>\%
    count_na(pivot = TRUE,
             drop_zeros = FALSE)
}
