% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project()
}
\description{
Creates common folder structure for new projects for consistency.
}
\examples{
\dontrun{
create_project()
}
}
