% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_compress.R
\name{csv_compress}
\alias{csv_compress}
\title{csv_compress}
\usage{
csv_compress(folder)
}
\arguments{
\item{folder}{Source / destination folder for compressing .csv files.}
}
\value{
Compressed files in the specified location.
}
\description{
Read a folder of .csv files and compress them individually to .gz
}
\examples{
\dontrun{
}

}
