% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{date_split}
\alias{date_split}
\title{date_split}
\usage{
date_split(x, col)
}
\arguments{
\item{x}{Dataframe}

\item{col}{Date format column to split}
}
\value{
Dataframe with added year, month, day, and yday columns.
}
\description{
Splits a date column into separate year, month, day, and julian day columns.
}
\examples{
\dontrun{
}

}
