% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffmc.R
\name{dffmc}
\alias{dffmc}
\title{fwi_dffmc}
\usage{
dffmc(x)
}
\arguments{
\item{x}{Fine Fuel Moisture Content value.}
}
\value{
Dead fine fuel moisture content (%: 0 - 250%).
}
\description{
Quick conversion of the Canadian FWI System's Fine Fuel Moisture Code (FFMC) to dead fine fuel moisture content (%). Because I'm sick of looking this up again and again.
}
\examples{
\dontrun{

  dffmc(c(84, 92))

}
}
