% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\name{dominance}
\alias{dominance}
\title{dominance}
\usage{
dominance(
  mod,
  n_boot = 1000,
  method = c("lmg", "pmvd", "last", "first", "betasq", "pratt", "genizi", "car"),
  seed = 8675309,
  ...
)
}
\arguments{
\item{mod}{Regression model input.}

\item{n_boot}{Number of bootstrapped model samples to assess.}

\item{method}{Method for assessing variable dominance, passed to {relaimpo}.}

\item{seed}{Random seed.}
}
\value{
Dataframe with model method, bootstrapped sample number, predictor variable name, dominance values, and variable ranks.
}
\description{
Runs a bootstrapped dominance analysis (aka relative weighting analysis or Shapley regression) via the {relaimpo} package and returns a dataframe of all relative importance values and the associated ranks.
}
\examples{
\dontrun{
mtcars |>
  lm(mpg ~ cyl + wt + hp,
     data = _) |>
    dominance(n_boot = 500,
              method = "car")
}

}
