% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hull.R
\name{extract_hull}
\alias{extract_hull}
\title{extract_hull}
\usage{
extract_hull(
  df,
  x,
  y,
  group = NULL,
  drop_outliers = FALSE,
  output = c("df", "sf")
)
}
\arguments{
\item{df}{Dataframe to extract shapes from.}

\item{x}{Numeric variable.}

\item{y}{Numeric variable.}

\item{group}{Grouping variable for comparing shape of groups along x-y axes.}

\item{drop_outliers}{Logical noting to remove outlier percentiles (< 5% | > 95%) on both x and y variables.}

\item{output}{Character -- either 'df' or 'sf' -- to return regular dataframe for use with `geom_polygon()`, or sf spatial object for use with `geom_sf()`. Note that spatial output applies a smoothing process to the shapes.}
}
\value{
Dataframe of ellipses line shapes along NMDS1 and NMDS2.
}
\description{
Extracts the shape of data from dataframes. Originally crafted around multidimensional scaling, but works well with any dataframe.
}
\examples{
{
data(iris)
iris |>
  extract_hull(x = "Sepal.Width",
               y = "Sepal.Length",
               group = "Species",
               drop_outliers = TRUE,
               output = "sf") |>
    ggplot() +
    geom_sf(aes(col = Species),
            fill = NA) +
    theme_meow()
}

}
