% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{left}
\usage{
left(x, n = 4, pad = NULL)
}
\arguments{
\item{x}{Vector to extract from.}

\item{n}{Number of characters to extract starting from the left.}

\item{pad}{Character to pad string values with if the input vectors are less than n.}
}
\value{
Transformed vector of `n` length.
}
\description{
Simplified variant of `substr` imitating the Excel & Alteryx `LEFT` function.
}
\examples{
left(c('Example', 'E.g.', 'Eg'),
     n = 4,
     pad = ' ')

}
