% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_arrow.R
\name{mds_arrow}
\alias{mds_arrow}
\title{mds_arrow}
\usage{
mds_arrow(mds, name = "var")
}
\arguments{
\item{mds}{Multidimensional scaling object created through `vegan::metaMDS()`.}

\item{name}{Character value to name column of variable names.}
}
\value{
Dataframe of variable names and NMDS1 and NMDS2 values to describe the 2-dimensional impact of these variables from [0,0].
}
\description{
Pulls arrow details from multidimensional scaling object via `vegan::metaMDS()`. Note that these represent arrows in a typical multidimensional scaling plot *from* x = 0, y = 0.
}
