% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_ellipse.R
\name{mds_ellipse}
\alias{mds_ellipse}
\title{mds_ellipse}
\usage{
mds_ellipse(df, mds, group)
}
\arguments{
\item{df}{Dataframe used in the creation of an MDS analysis.}

\item{mds}{Multidimensional scaling object created through `vegan::metaMDS()`.}

\item{name}{Character column name representing the grouping variable we want to highlight.}
}
\value{
Dataframe of ellipses line shapes along NMDS1 and NMDS2.
}
\description{
Pulls ellipsis shapes from multidimensional scaling outputs.
}
