% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_env.R
\name{mod_env}
\alias{mod_env}
\title{mod_env}
\usage{
mod_env(quiet = TRUE)
}
\arguments{
\item{quiet}{Logical. Whether to print the changed environment paths or not.}
}
\value{
Modified internal `HOME` and `R_USER` paths.
}
\description{
Modify internal paths for `HOME` and `R_USER` options, removing potential issues with OneDrive defaults on Windows 11. Stands for "Microsoft OneDrive Environments."
}
\examples{
\dontrun{

   mod_env()

}
}
