% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_covariates.R
\name{model_covariates}
\alias{model_covariates}
\title{model_covariates}
\usage{
model_covariates(mod, plot_rows = 1, plot_vif = FALSE, ...)
}
\arguments{
\item{mod}{Input model to assess.}
}
\value{
Output from tidy_vif and figures of dominance values.
}
\description{
Uses both tidy_vif and dominance to assessment model covariates. Shorthand for plotting dominance values and variable ranks, while also printing variance inflation factor scores.
}
\examples{
\dontrun{

  mtcars |>
    lm(mpg ~ cyl + disp + hp + wt,
       data = _) |>
    model_covariates(plot_rows = 2,
                     plot_vif = TRUE,
                     n_boot = 500,
                     method = "car")

}
}
