% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulo.R
\name{modulo}
\alias{modulo}
\title{modulo}
\usage{
modulo(x, res = 1)
}
\arguments{
\item{x}{Numeric vector to scale.}

\item{res}{Base resolution to scale x to.}
}
\value{
Scaled vector.
}
\description{
Applies a modulo operation with a specified base resolution
}
\examples{
runif(100, 0, 100) \%>\%
 tibble::as_tibble() \%>\%
 dplyr::mutate(mod = modulo(value, 15))

}
