% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rate.R
\name{na_rate}
\alias{na_rate}
\title{na_rate}
\usage{
na_rate(x, rate = 1.1)
}
\arguments{
\item{x}{Vector with NAs in future time to fill}
}
\value{
Numeric values with a constant rate applied to lagged values.
}
\description{
Extend a timeseries dataset and apply an annual growth rate to it
}
\examples{
\dontrun{
tibble::tibble(START_YEAR = c(2018:2020),
                     VALUE = c(1,2,3)) \%>\%
 dplyr::bind_rows(tibble::tibble(START_YEAR = c(2021:2024))) \%>\%
 dplyr::mutate(VALUE = na_rate(VALUE, rate = 0.98))

}

}
