% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvl.R
\name{nvl}
\alias{nvl}
\title{nvl}
\usage{
nvl(x, value = 0)
}
\arguments{
\item{x}{Vector to clean of any class.}

\item{value}{Replacement value to substitute.}
}
\value{
Cleaned vector.
}
\description{
Replaces all no-value options with a specified replacement.
}
\examples{
nvl(c(NA, NaN, Inf, -Inf, NA_character_, NA_real_, NA_complex_, NA_integer_), 0)

}
