% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_indices.R
\name{order_indices}
\alias{order_indices}
\title{order_indices}
\usage{
order_indices(x)
}
\arguments{
\item{x}{Vector to create group indices from.}
}
\value{
Numeric indices for input group variable.
}
\description{
Tweak of `dplyr`'s group_indices allowing for retaining the original order.
}
\examples{
\dontrun{

}

}
