% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{package_check}
\alias{package_check}
\title{package_check}
\usage{
package_check(
  packages,
  quiet = FALSE,
  repos = c("https://cloud.r-project.org")
)
}
\arguments{
\item{packages}{chr : Single character or character vector with package/library names}

\item{quiet}{lgl : Boolean whether to show which packages were installed, loaded, or skipped, as well as their version numbers.}

\item{repos}{chr : Specified repo web address to attempt to pull uninstalled libraries from.}
}
\value{
Installs and/or loads libraries on the backend.
}
\description{
Install and/or load any necessary libraries, or skip if they are already loaded.
}
\examples{
\dontrun{
packacge_check(c('tidyverse', 'utastoolkit'))
}
}
