% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_sf.R
\name{pal_sf}
\alias{pal_sf}
\title{New Ace Science Fiction Series 3 colour palettes}
\usage{
pal_sf(book = "neuromancer", n = 3, type = "div")
}
\arguments{
\item{book}{Reference identifier to one of the 12 books.}

\item{n}{Number of colors to return.}

\item{type}{Type of colour palette, either sequential or diverging.}
}
\value{
Vector of colour hex codes.
}
\description{
`pal_sf` produces sequential or divergent colour palettes
based on the 12 original book covers released in the third
Ace SF series released between 1984 and 1990.
}
\examples{
\dontrun{
 pal_sf(book = 'green eyes',
        n = 6,
        type = 'div')
}
}
