% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{pkgs}
\alias{pkgs}
\title{pkgs}
\usage{
pkgs(packages, quiet = FALSE, repos = c("https://cloud.r-project.org"))
}
\arguments{
\item{packages}{Single character or character vector with package/library names}

\item{quiet}{Logical whether to show which packages were installed, loaded, or skipped, as well as their version numbers.}

\item{repos}{Specified repo web address to attempt to pull uninstalled libraries from.}
}
\value{
Environmental conditions: Installs and/or loads libraries where necessary.
}
\description{
Install and/or load any necessary libraries, or skip if they are already loaded.
}
\examples{
\dontrun{
miao::pkgs(c('toddellis/miao', 'tidyverse'))
}
}
