% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprint.R
\name{pprint}
\alias{pprint}
\title{pprint}
\usage{
pprint(
  p,
  file = NULL,
  dpi = 540,
  overwrite = FALSE,
  folder = "02 Results/Figures",
  ...
)
}
\arguments{
\item{p}{Figure to print.}

\item{file}{Output figure name.}

\item{dpi}{Dots per inch.}

\item{overwrite}{Logical to replace existing file or not if it exists.}

\item{folder}{Output figure subfolder (assuming standardised project folder structure).}
}
\description{
Shortcut to print out .png and .pdf figures at a high resolution to project subfolder.
}
\examples{
\dontrun{


}

}
