% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrrallel.R
\name{purrrallel}
\alias{purrrallel}
\title{purrrallel}
\usage{
purrrallel(x, nclust = NULL, seed = 8675309)
}
\arguments{
\item{x}{Piped workflow or dataset with some methods that could use parallel processing.}

\item{nclust}{Number of clusters to set.}
}
\value{
Function output with number of clusters applied and then flushed.
}
\description{
Set parallel processing clusters within a workflow.
}
