% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arrow.R
\name{read_arrow}
\alias{read_arrow}
\title{read_arrow}
\usage{
read_arrow(files)
}
\arguments{
\item{files}{Vector of file paths to be read by}
}
\value{
Bound dataset comprised of multiple files.
}
\description{
Shorthand for {arrow}'s confusing syntax to read a directory.
}
\examples{
\dontrun{
}

}
