% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs.R
\name{read_fs}
\alias{read_fs}
\alias{read_folder}
\title{read_folder}
\usage{
read_fs(files, colClasses = c(NULL), method = c("purrr", "base"))

read_folder(files, colClasses = c(NULL), method = c("purrr", "base"))
}
\arguments{
\item{files}{Filelist generated using fs::dir_ls.}

\item{colClasses}{Column classes if desired.}

\item{method}{Option to use for loop or purrr::map.}
}
\value{
Either bound dataframe (via loop) or nested dataset (via purrr).
}
\description{
Loop through objects in an fs::dir_ls filelist and apply data.table::fread across them.
}
\examples{
\dontrun{
}

}
