% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_yday.R
\name{scale_x_yday}
\alias{scale_x_yday}
\title{scale_x_yday}
\usage{
scale_x_yday()
}
\value{
ggplot figure with monthly breaks and labels on x axis.
}
\description{
Converts lubridate::yday (Julian day of the year) axis to breaks and labels noting month of the year
}
