% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_area.R
\name{sf_area}
\alias{sf_area}
\title{sf_area}
\usage{
sf_area(x)
}
\arguments{
\item{x}{sf feature type.}
}
\value{
Dataframe with added column representing area
}
\description{
Extract area values from sf features.
}
\examples{
\dontrun{
}

}
