% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_coords.R
\name{sf_coords}
\alias{sf_coords}
\title{sf_coords}
\usage{
sf_coords(x)
}
\arguments{
\item{x}{sf feature type.}
}
\value{

}
\description{
Extract x-y coordinate values from sf features.
}
\examples{
\dontrun{
}

}
