% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_dims.R
\name{sf_dims}
\alias{sf_dims}
\title{sf_dims}
\usage{
sf_dims(x)
}
\arguments{
\item{x}{sf feature type.}
}
\value{
Dataframe with added columns representing area, length, and perimeter values
}
\description{
Returns area, length, and perimeter for vector features.
}
\examples{
\dontrun{
}

}
