% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{subsample}
\usage{
subsample(x, rows_n = 100, samples_n = 100, replace = TRUE)
}
\arguments{
\item{x}{Dataset}

\item{rows_n}{Number of rows to sample from the dataset}

\item{samples_n}{Number of times to sample the dataset}
}
\value{
Dataframe
}
\description{
Combines purrr and dplyr sampling methods to sample x rows from a dataset y times. Useful for means testing.
}
\examples{
\dontrun{

}

}
