% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_sentiment.R
\name{summarise_sentiment}
\alias{summarise_sentiment}
\alias{summarize_sentiment}
\title{summarise_sentiment}
\usage{
summarise_sentiment(
  data,
  x,
  line_length = 0,
  lex_stop = c("all", "none", "onix", "smart", "snowball")
)

summarize_sentiment(
  data,
  x,
  line_length = 0,
  lex_stop = c("all", "none", "onix", "smart", "snowball")
)
}
\arguments{
\item{data}{dataframe object}

\item{x}{text column to pull sentiment from}
}
\value{
summary sentiment scores for all available sentiment lexicon dictionaries.
}
\description{
Summarises sentiment for all sentiment dictionaries.
}
\examples{
\dontrun{
janeaustenr::austen_books() |>
  dplyr::group_by(book) |>
  summarise_sentiment(text)
}

}
