% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_monkey.R
\name{survey_monkey}
\alias{survey_monkey}
\title{survey_monkey}
\usage{
survey_monkey(survey, pivot = FALSE)
}
\arguments{
\item{pivot}{Logical determinant of whether to pivot the dataset from long format (default) to wide format data.}

\item{x}{Dataframe of raw Survey Monkey dataset.}
}
\value{
Dataframe of cleaned (or at least more manageable) Survey Monkey data.
}
\description{
Clean (most) Survey Monkey format data, barring overly-complex and poorly-constructed surveys (insert tears).
}
\examples{
\dontrun{

}

}
