% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_explorer.R
\name{tbl_explorer}
\alias{tbl_explorer}
\title{tbl_explorer}
\usage{
tbl_explorer(x, min = 1, drop_numeric = FALSE)
}
\arguments{
\item{x}{Dataset}

\item{min}{Numeric cutoff point for minimum number of observation occurrences. Can be a proportion or an integer value.}

\item{drop_numeric}{Logical argument to drop numeric columns with a lot of variability}
}
\value{
A dataframe with column and value counts.
}
\description{
Counts all unique values in an entire dataset across all columns
}
\examples{
utas_tbl('students.campuses') \%>\%
  tbl_exporer()

}
