% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_na.R
\name{tbl_na}
\alias{tbl_na}
\title{tbl_na}
\usage{
tbl_na(x, pivot = TRUE, drop_zeros = TRUE)
}
\arguments{
\item{x}{Dataframe.}

\item{pivot}{Logical whether to pivot the output to long data format.}

\item{drop_zeros}{Logical whether to drop groups or columns with zero missing values.}
}
\value{
Wide or long format dataframe with the count of zeros across all the columns in the original dataset.
}
\description{
Check the number of NAs for all columns and provide a summary view.
}
\examples{
tibble::tibble(a = c(1, 2, NA),
               b = c(NA, NA, 'Q'),
               c = c(3, 2, 1)) \%>\%
    count_na(pivot = TRUE,
             drop_zeros = FALSE)
}
