% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary.R
\name{tbl_summary}
\alias{tbl_summary}
\title{tbl_summary}
\usage{
tbl_summary(x, .prop = FALSE)
}
\arguments{
\item{x}{Dataset.}

\item{.prop}{Logical determining whether output should transform the count of distinct and total observations for each column into a proportion of the dataset's total size.}
}
\value{
A dataframe with the number of distinct observations and total observations per column.
}
\description{
Gets the number of distinct values in a dataset and total number of observations per column.
}
\examples{
tibble::tibble(a = c('A','B', NA),
               b = c(1.2, 1.2, 1.2),
               c = c(4.5, NA, NA)) \%>\%
 tbl_summary(.prop = TRUE)

}
