% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gam.R
\name{tidy_gam}
\alias{tidy_gam}
\title{tidy_gam}
\usage{
tidy_gam(x, conf.level = 0.95)
}
\arguments{
\item{x}{GAM model object}

\item{conf.level}{Confidence level for pulling confidence intervals from parametric terms.}
}
\value{
Dataframe with model/term metrics.
}
\description{
Wrapped for 2x broom::tidy as applied to a GAM object, pulling both parametric and smooth terms.
}
\examples{
\dontrun{
}


}
