% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_vif.R
\name{tidy_vif}
\alias{tidy_vif}
\title{tidy_vif}
\usage{
tidy_vif(mod)
}
\arguments{
\item{mod}{Input model to assess for variance inflation among predictors.}
}
\value{
Dataframe with model/term metrics.
}
\description{
Quick and tidy-friendly adaptation of {car}'s `vif()`. Returns a dataframe with variance inflation, tolerance, and a flag identifying if either is beyond common cut-offs reflecting problematic input variables.
}
\examples{
\dontrun{
}


}
