% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_stats.R
\name{trend_stats}
\alias{trend_stats}
\title{trend_stats}
\usage{
trend_stats(
  x,
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  continuity = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{x}{dbl : numeric timeseries variable}

\item{conf_level}{dbl : confidence level for upper- and lower-bounds for the Sen-Theil slope, defaults to 0.95 (p = 0.05)}

\item{alternative}{chr : one of three hypotheses to apply to Mann-Kendall statistics for testing trend direction}

\item{continuity}{lgl : TRUE or FALSE on whether to apply continuity correction to Mann-Kendall statistics. Same correction is always applied to Sen-Theil slope estimation.}

\item{na.rm}{lgl : TRUE or FALSE on whether to remove missing values. If missing values exist and this is FALSE, function will stop running.}
}
\value{
.bsen dbl : Sen-Theil slope estimator

.bsen_lwr dbl : Lower bounds for Sen-Theil slope estimator

.bsen_upr dbl : Upper bounds for Sen-Theil slope estimator

.mks dbl : Mann-Kendall S statistic

.mks_var dbl : Mann-Kendall trend variability

.mkz dbl : Mann-Kendall Z statistic

.ktau : Kendall's tau

.ktau_d : Kendall's tau D statistic

.p dbl : p-value for all trend statistics. Dependent on `continuity = TRUE`

.bsen_p dbl : p-value for Sen-Theil slope estimator and Kendall's tau(?). Dependent on `continuity = FALSE`

.mkz_p dbl : p-value for Mann-Kendall statistics. Dependent on `continuity = FALSE`
}
\description{
Applies a family of robust trend statistics: The Mann-Kendall S statistic, the Sen-Theil slope estimator, and Kendall's tau.
}
\examples{
\dontrun{

library(tidyverse)
dummy <- tibble::tibble(YEAR = c(2000:2020)) \%>\%
  rowwise() \%>\%
  mutate(RANDOM = runif(1, -33, 100)) \%>\%
  ungroup() \%>\%
  mutate(GRP = 'GROUP A') \%>\%
  bind_rows(tibble::tibble(YEAR = c(2005:2020)) \%>\%
              rowwise() \%>\%
              mutate(RANDOM = runif(1, 33, 100)) \%>\%
              ungroup() \%>\%
              mutate(GRP = 'GROUP B')) \%>\%
  bind_rows(tibble::tibble(YEAR = c(2000:2020)) \%>\%
              rowwise() \%>\%
              mutate(RANDOM = runif(1, 0, 1)) \%>\%
              ungroup() \%>\%
              mutate(GRP = 'GROUP C'))
 dummy \%>\%
   group_by(GRP) \%>\%
   summarise(trend_stats(RANDOM))

}

}
