% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegan_scores.R
\name{vegan_scores}
\alias{vegan_scores}
\title{vegan_scores}
\usage{
vegan_scores(df, mds)
}
\arguments{
\item{df}{Dataframe used in {vegan} MDS analyses. N.B. Should add other methods like PCA later.}

\item{mds}{Multidimensional scaling object created using `vegan::metaMDS()`.}
}
\value{
Dataframe with NMDS1 and NMDS2 added as columns.
}
\description{
Adds scores from {vegan} objects to initial dataset.
}
