% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcount.R
\name{wordcount}
\alias{wordcount}
\title{wordcount}
\usage{
wordcount(x)
}
\arguments{
\item{x}{String variable or vector.}
}
\value{
Wordcount variable or vector.
}
\description{
Count the number of distinct words in a sentence.
}
\examples{
\dontrun{
}

}
