% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{zero_one}
\alias{zero_one}
\title{zero_one}
\usage{
zero_one(x)
}
\arguments{
\item{x}{Vector to transform.}
}
\value{
Numeric values ranging between 0 and 1.
}
\description{
Scale a numeric vector to between 0 and 1. N.B. Kind of pointless given `scales::rescale` exists.
}
\examples{
\dontrun{
tibble::tibble(YEAR = c(2018:2020),
               VALUE = c(1,2,3)) \%>\%
 dplyr::mutate(YEAR_SCALED = zero_one(YEAR),
               VALUE_SCALED = zero_one(VALUE))

}
}
