% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectable.R
\name{collectable}
\alias{collectable}
\title{collectable}
\usage{
collectable(x, qid = "query", format = "tibble")
}
\arguments{
\item{x}{Dataset or lazy database query.}

\item{qid}{Optional string query identifier.}

\item{format}{Output format type, either via `data.table`, `dtplyr`, or `tibble`.}
}
\value{
A queried database connection stored in local memory and the time it took to pull into memory. Alternatively provides the time to process a piped workflow or single function if the input dataset is not a lazy database connection.
}
\description{
Wraps a timer around functions that pulls a lazy database query into memory.
}
\examples{
\dontrun{
dplyr::tbl(connection,
           dbplyr::in_schema("SCHEMA",
                             "TABLE_NAME")) \%>\%
  collectable(qid = "an impossible and certainly fake query",
              format = 'tibble')
}

}
