% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_area.R, R/sf_coords.R
\name{sf_coords}
\alias{sf_coords}
\title{sf_area}
\usage{
sf_coords(x)

sf_coords(x)
}
\arguments{
\item{x}{sf feature type.}
}
\value{
Dataframe with added column representing area

Dataframe with added columns representing centroid x-y coordinates.
}
\description{
Extract area values from sf features.

Extract x-y coordinate values from sf features.
}
\examples{
\dontrun{
}

\dontrun{
}

}
